common: Move the event enums into a separate module
